My ETL functions

My ETL functions

# S3 method for etl_atmoschem.process
etl_init(
  obj,
  script = NULL,
  schema_name = "init",
  pkg = attr(obj, "pkg"),
  ext = NULL,
  ...
)

Arguments

obj
An etl object

script
either a vector of SQL commands to be executed, or a file path as a character vector containing an SQL initialization script. If NULL (the default), then the appropriate built-in schema will be fetched by find_schema, if it exists. Note that the flavor of SQL in this file must match the type of the source. That is, if your object is of type src_mysql, then make sure that the schema you specify here is written in MySQL (and not PostgreSQL). Please note that SQL syntax is not, in general, completely portable. Use with caution, as this may clobber any existing data you have in an existing database.

schema_name
The name of the schema. Default is init.

pkg
The package defining the schema. Should be set in etl.

ext
The file extension used for the SQL schema file. If NULL (the default) it be inferred from the src_* class of con. For example, if con has class src_sqlite then ext will be sqlite.

...
Currently ignored

Examples

if (FALSE) {
if (require(dplyr)) {
  obj <- etl("atmoschem.process") %>%
    etl_create()
}
}