Flag aggregated data

Return the NARSTO flag for an aggregated measurement (for example, hourly average or maximum). See NARSTO (2001, p. 10) .

narsto_agg_flag(p, below_mdl = FALSE)

Arguments

p

Proportion of valid measurements.

below_mdl

Whether the aggregated value is below the method detection limit. MDLs for aggregates are generally lower than the limits for individual measurements.

Value

A NARSTO flag.

References

NARSTO (2001). “Consensus Metadata Standard: Data Quality Flags.” https://cdiac.ess-dive.lbl.gov/programs/NARSTO/metadatastandards/consensus_flag_standard_20011105.pdf.

Examples

narsto_agg_flag(.7, below_mdl = FALSE)
#> [1] "V4"