Rolling median absolute deviation (MAD)

Calculate the median absolute deviation (MAD) of a rolling window of values.

rolling_mad(x, k)

Arguments

x
A vector of numbers.

k
Width of the rolling window (an odd integer).

Value

A vector of MADs.

Details

The median absolute deviation (MAD) is a median-based alternative to the standard deviation often preferred for its robustness to outliers (Leys et al. 2013) . The MAD for a collection of values \(X\) is defined as

$$1.4826 \times \textrm{med}(\left| X - \textrm{med}(X) \right|)$$

where the constant 1.4826 scales the result to estimate the standard deviation.

References

Leys C, Ley C, Klein O, Bernard P, Licata L (2013). “Detecting outliers: Do not use standard deviation around the mean, use absolute deviation around the median.” Journal of Experimental Social Psychology, 49(4), 764--766. ISSN 0022-1031, doi: 10.1016/j.jesp.2013.03.013 , http://www.sciencedirect.com/science/article/pii/S0022103113000668.

Examples

rolling_mad(rnorm(20), 5)

#>  [1] 0.07798711 0.27785782 0.39974143 0.68258364 1.24301517 1.11747511
#>  [7] 1.08132519 1.01648642 1.01648642 1.20214937 1.12510745 1.02548140
#> [13] 0.19518041 0.09962606 0.36477163 0.36477163 0.21533316 0.21533316
#> [19] 0.02535894 1.02355283