Sea level pressure

Estimate sea level pressure from local pressure, temperature, and height using the barometric formula.

sea_level_pressure(press, temp, height)

Arguments

press
Pressure in millibars.

temp
Temperature in Celsius.

height
Height above sea level in meters.

Value

Estimated sea level pressure.

Details

We estimate the sea level pressure using the barometric formula,

$$P = P_0 (\frac{T_0}{T})^\frac{Mg}{R^*L}$$

where \(P\) is pressure, \(T\) is temperature (in Kelvin), \(H\) is height, and \(L\) is the temperature lapse rate. Variables with subscript 0 are the same at sea level. Substituting \(T_0 = T - LH\), we can solve for \(P_0\) in terms of pressure, temperature, and height:

$$P_0 = P (\frac{T}{T - LH})^\frac{Mg}{R^*L}$$

The values of physical constants are taken from COESA (1976) :

\(L\) (lapse rate)

-.0065 K/m

\(M\) (molar mass of air)

0.0289644 kg/mol

\(g\) (gravitational acceleration)

9.80665 m/s^2

\(R^*\) (universal gas constant)

8.31432 J/(mol K)

References

COESA (1976). “U.S. Standard Atmosphere, 1976.” https://ntrs.nasa.gov/citations/19770009539.

Examples

# estimate sea level pressure at the Whiteface Mountain summit
sea_level_pressure(800, 30, 1483.5)

#> [1] 943.1199