Wood smoke indicator based on light absorption

Derive a semi-quantitative wood smoke indicator from aethalometer measurements using the Delta-C method.

wood_smoke(bc370, bc880)

Arguments

bc370

370nm estimate of black carbon (UVBC).

bc880

880nm estimate of black carbon.

Value

Delta-C wood smoke indicator values.

Details

Wood smoke is estimated using the Delta-C method introduced by Allen et al. (2004) . Delta-C (sometimes called UVPM) is the difference between aethalometer estimates at 370nm (sometimes called UVBC) and 880nm:

$$Delta\textrm{-}C = BC_{370nm} - BC_{880nm}$$

While absorption in the range of 880nm is known to identify black carbon, additional material absorbs light in the ultraviolet range measured at 370nm. Specifically, organic compounds associated with wood smoke become more absorbent and are included in the 370nm measurement.

The resulting estimate is "semi-quantitative" in the sense that it doesn't represent the true quantity of wood smoke particles-- rather it is roughly proportional to the true value.

References

Allen GA, Babich P, Poirot RL (2004). “Evaluation of a New Approach for Real Time Assessment of Wood Smoke PM.” https://www.nescaum.org/documents/2004-10-25-allen-realtime_woodsmoke_indicator_awma.pdf.

Wang Y, Hopke PK, Rattigan OV, Zhu Y (2011). “Characterization of ambient black carbon and wood burning particles in two urban areas.” Journal of Environmental Monitoring, 13(7), 1919--1926. ISSN 1464-0333, doi:10.1039/C1EM10117J , Publisher: The Royal Society of Chemistry, https://pubs.rsc.org/en/content/articlelanding/2011/em/c1em10117j.

Zhang KM, Allen G, Yang B, Chen G, Gu J, Schwab J, Felton D, Rattigan O (2017). “Joint measurements of PM\(_{\textrm{2.5}}\) and light-absorptive PM in woodsmoke-dominated ambient and plume environments.” Atmospheric Chemistry and Physics, 17(18), 11441--11452. ISSN 1680-7316, doi:10.5194/acp-17-11441-2017 , https://www.atmos-chem-phys.net/17/11441/2017/.

Examples

wood_smoke(1000, 800)
#> [1] 200